requirements.txt
Markdown==3.2.2
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==1.0
psycopg2-binary==2.8.5
Flask-Jsonpify==1.5.0
Flask-JWT==0.3.2
